--- routes/retrieveRoutes.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const retrieveController = require('../controllers/retrieveController');
-
-
-router.route('/profilePicture')
-    .get(retrieveController.getProfilePicture)
-
-module.exports = router
